# docs/04-API명세-openapi.yaml
openapi: 3.0.3
info:
  title: HOOPLY API
  version: 1.0.0
  description: 농구 동호회 팀 매칭 서비스. 인증은 Bearer JWT(access 30분/refresh 14일). 모든 4xx/5xx 본문은 ErrorResponse {code, message, details[]}.
servers:
- url: https://hooply-backend.onrender.com/api/v1
  description: 운영
- url: http://localhost:8000/api/v1
  description: 로컬
security:
- bearerAuth: []
paths:
...
  /teams/{team_id}/guests:
    post:
      summary: 게스트 레코드 등록
      tags:
      - 게스트 관리
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestCreate'
      responses:
        '201':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerCard'
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestSimilar'
    get:
      summary: 기존 게스트 검색
      tags:
      - 게스트 관리
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      - name: q
        in: query
        required: false
        schema:
          type: string
          nullable: true
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemList_PlayerCard_'
  /teams/{team_id}/guests/merge-candidates:
    get:
      summary: 게스트–회원 병합 후보
      tags:
      - 게스트 관리
      parameters:
      - name: team_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemList_MergeCandidate_'
        '403':
          description: FORBIDDEN_ROLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /players/{player_id}:
    patch:
      summary: 게스트 정보 수정
      tags:
      - 게스트 관리
      parameters:
      - name: player_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerUpdate'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerCard'
        '403':
          description: FORBIDDEN_NOT_OWNER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /players/{guest_player_id}:merge:
    post:
      summary: 게스트를 회원 계정에 병합
      tags:
      - 게스트 관리
      parameters:
      - name: guest_player_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerCard'
        '403':
          description: FORBIDDEN_ROLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: ALREADY_MERGED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: MERGE_KIND_MISMATCH
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /players/{player_id}:unmerge:
    post:
      summary: 게스트 병합 되돌리기
      tags:
      - 게스트 관리
      parameters:
      - name: player_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerCard'
        '403':
          description: FORBIDDEN_ROLE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /players/{guest_player_id}:claim:
    post:
      summary: 게스트 기록 본인 확인 (병합 / 거절)
      tags:
      - 게스트 관리
      parameters:
      - name: guest_player_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuestClaimIn'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemList_GuestClaimView_'
        '403':
          description: FORBIDDEN_ROLE / NOT_A_MEMBER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: ALREADY_MERGED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /events/{event_id}/guests:
    post:
      summary: 회차에 게스트 등록
      tags:
      - 일정 · 참석
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventGuestCreate'
      responses:
        '201':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceView'
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestSimilar'
  /events/{event_id}/guests/{player_id}:
    patch:
      summary: 회차 게스트 수정
      tags:
      - 일정 · 참석
      parameters:
      - name: player_id
        in: path
        required: true
        schema:
          type: integer
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventGuestUpdate'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceView'
        '403':
          description: FORBIDDEN_NOT_OWNER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: 회차 게스트 삭제 (불참 처리)
      tags:
      - 일정 · 참석
      parameters:
      - name: player_id
        in: path
        required: true
        schema:
          type: integer
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          description: 성공
        '403':
          description: FORBIDDEN_NOT_OWNER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AttendanceStatus:
      type: string
      enum:
      - ATTEND
      - ABSENT
      - PENDING
...
    AttendanceView:
      properties:
        player:
          $ref: '#/components/schemas/PlayerCard'
        status:
          $ref: '#/components/schemas/AttendanceStatus'
        note:
          type: string
          nullable: true
        responded_at:
          type: string
          format: date-time
          nullable: true
        registered_by:
          type: integer
          nullable: true
        registered_by_name:
          type: string
          nullable: true
        team_lock_request_player_id:
          type: integer
          nullable: true
        team_lock_request_player_name:
          type: string
          nullable: true
        can_edit:
          type: boolean
      type: object
      required:
      - player
      - status
...
    ErrorDetail:
      properties:
        field:
          type: string
          nullable: true
        reason:
          type: string
        context:
          additionalProperties: true
          type: object
          nullable: true
      type: object
      required:
      - reason
    ErrorResponse:
      properties:
        code:
          type: string
        message:
          type: string
        details:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
      type: object
      required:
      - code
      - message
...
    EventGuestCreate:
      properties:
        display_name:
          type: string
          maxLength: 50
          minLength: 1
        skill_grade:
          type: integer
          maximum: 5.0
          minimum: 1.0
          nullable: true
        height_cm:
          type: integer
          maximum: 250.0
          minimum: 120.0
          nullable: true
        preferred_position:
          $ref: '#/components/schemas/Position'
          nullable: true
        playable_positions:
          items:
            $ref: '#/components/schemas/Position'
          type: array
        team_lock_request:
          type: boolean
        existing_player_id:
          type: integer
          nullable: true
        force_new:
          type: boolean
      type: object
      required:
      - display_name
    EventGuestUpdate:
      properties:
        display_name:
          type: string
          maxLength: 50
          minLength: 1
          nullable: true
        skill_grade:
          type: integer
          maximum: 5.0
          minimum: 1.0
          nullable: true
        height_cm:
          type: integer
          maximum: 250.0
          minimum: 120.0
          nullable: true
        preferred_position:
          $ref: '#/components/schemas/Position'
          nullable: true
        playable_positions:
          items:
            $ref: '#/components/schemas/Position'
          type: array
          nullable: true
        team_lock_request:
          type: boolean
          nullable: true
        status:
          $ref: '#/components/schemas/AttendanceStatus'
          nullable: true
      type: object
...
    GuestClaimIn:
      properties:
        accept:
          type: boolean
      type: object
      required:
      - accept
    GuestClaimView:
      properties:
        guest:
          $ref: '#/components/schemas/PlayerCard'
        team_id:
          type: integer
        team_name:
          type: string
        member_player_id:
          type: integer
        events_attended:
          type: integer
        quarters_played:
          type: integer
        last_event_date:
          type: string
          format: date
          nullable: true
      type: object
      required:
      - guest
      - team_id
      - team_name
      - member_player_id
      - events_attended
      - quarters_played
    GuestCreate:
      properties:
        display_name:
          type: string
          maxLength: 50
          minLength: 1
        skill_grade:
          type: integer
          maximum: 5.0
          minimum: 1.0
          nullable: true
        height_cm:
          type: integer
          maximum: 250.0
          minimum: 120.0
          nullable: true
        preferred_position:
          $ref: '#/components/schemas/Position'
          nullable: true
        playable_positions:
          items:
            $ref: '#/components/schemas/Position'
          type: array
        force_new:
          type: boolean
      type: object
      required:
      - display_name
...
    GuestSimilar:
      properties:
        similar:
          items:
            $ref: '#/components/schemas/PlayerCard'
          type: array
      type: object
      required:
      - similar
...
    ItemList_GuestClaimView_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/GuestClaimView'
          type: array
      type: object
      required:
      - items
...
    ItemList_MergeCandidate_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MergeCandidate'
          type: array
      type: object
      required:
      - items
    ItemList_PlayerCard_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PlayerCard'
          type: array
      type: object
      required:
      - items
...
    MergeCandidate:
      properties:
        guest:
          $ref: '#/components/schemas/PlayerCard'
        member:
          $ref: '#/components/schemas/PlayerCard'
      type: object
      required:
      - guest
      - member
    MergeRequest:
      properties:
        into_player_id:
          type: integer
      type: object
      required:
      - into_player_id
...
    PlayerCard:
      properties:
        id:
          type: integer
        user_id:
          type: integer
          nullable: true
        kind:
          type: string
        display_name:
          type: string
        role:
          type: string
        profile_image_url:
          type: string
          nullable: true
        height_cm:
          type: integer
          nullable: true
        skill_grade:
          $ref: '#/components/schemas/SkillGrade'
          nullable: true
        primary_position:
          $ref: '#/components/schemas/Position'
          nullable: true
        playable_positions:
          items:
            $ref: '#/components/schemas/Position'
          type: array
        attendance_rate:
          type: number
          nullable: true
        skill_confidence:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          nullable: true
      type: object
      required:
      - id
      - kind
      - display_name
      - role
...
    PlayerUpdate:
      properties:
        display_name:
          type: string
          maxLength: 50
          minLength: 1
          nullable: true
        skill_grade:
          type: integer
          maximum: 5.0
          minimum: 1.0
          nullable: true
        height_cm:
          type: integer
          maximum: 250.0
          minimum: 120.0
          nullable: true
        preferred_position:
          $ref: '#/components/schemas/Position'
          nullable: true
        playable_positions:
          items:
            $ref: '#/components/schemas/Position'
          type: array
          nullable: true
      type: object
...
    Position:
      type: string
      enum:
      - PG
      - SG
      - SF
      - PF
      - C
...
    SkillGrade:
      type: string
      enum:
      - A
      - B
      - C
      - D
      - E
...